# Conflicts:
#	frontend/src/app/+form/form-info/form-info.component.ts
#	frontend/src/app/+form/form-info/publish-pending/publish-pending.component.ts
#	frontend/src/app/formly/types/uvp-sections/uvp-sections.component.html

+++ server/src/main/java/de/ingrid/igeserver/exports/iso/Decimal.kt
@@ -23,5 +23,5 @@ import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty
 
 data class Decimal(
     @JacksonXmlProperty(localName = "Decimal", namespace = "http://www.isotc211.org/2005/gco")
-    var value: Float? = null,
+    var value: Double? = null,
 )
